export_keys/requirements.txt
ton-client-py==1.30.0.0
base58==1.0.3
